Finally upload current progress.
Initial upload of progress thus far.
Project is nearing preparation for release alongside it's other 3 dependent mods (rhynia_leiodora, rhynia_canaria and rhynia_mosa). Still, VERY insecure code, plagued with inconsistencies and redundancies in aspects such as function hooking, data transfer, data acquisition, etc; nevertheless, uploading at this stage was a priority for a variety of reasons. Other content and ideas from the old "$nc_flora" repository are scheduled to be incorporated as either additions/revisions to this mod, or as "rhynia_" module mods in like manner to the aforementioned 3 mods.