Branch Commit 1, fixed naming, mod.conf
This branch tracks the changes to Rhynia that will be made under the strategy of code simplification. Much of the main Rhynia files are a haphazard mess of repetitive, poorly communicating, and at times spaghettified code that was made during a time of significant constraint on developer time. Reading over it with an aim to improve, several key areas spanning both current and planned architecture could benefit from significant simplification. This would likely increase both the efficiency, readability, cohesiveness and cleanliness of the existing code; while also making future development and deadline-satisfaction more easily achievable.